# api/openapi.yaml
openapi: 3.1.0
info:
  title: Blade Runner
  description: Blade Runner desc
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development

paths:
  /github-api/repos:
    get:
      summary: Get user's GitHub repositories
      operationId: getMyRepos
      security:
        - githubOAuth: [read:user, repo]
      tags:
        - Github
      responses:
        '200':
          description: A list of the user's GitHub repositories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GithubRepo'

  /github-api/user:
    get:
      summary: Get authenticated user information
      operationId: getUserInfo
      security:
        - githubOAuth: [read:user]
      tags:
        - Github
      responses:
        '200':
          description: Details of the currently authenticated user
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true

  /github-api/diff:
    get:
      summary: Retrieve Pull Request Diff
      description: Fetches the raw diff string for a specific Pull Request using the GitHub API.
      operationId: getPullRequestDiff
      tags:
        - Github
      parameters:
        - name: owner
          in: query
          required: true
          description: The GitHub repository owner (username or organization name).
          schema:
            type: string
            example: "octocat"
        - name: repository
          in: query
          required: true
          description: The name of the GitHub repository.
          schema:
            type: string
            example: "Hello-World"
        - name: pullRequestNumber
          in: query
          required: true
          description: The pull request number/ID.
          schema:
            type: string
            example: "123"

      responses:
        '200':
          description: Successfully retrieved SVG diff representation.
          content:
            image/svg+xml:
              schema:
                type: string
                format: binary
                description: The raw SVG XML content.
        '400':
          description: Bad Request - Missing or invalid query parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found - The repository or pull request does not exist, or the backend lacks permissions to access it.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error - Backend failed to communicate with the upstream GitHub API.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /pr-metadata/files:
    get:
      summary: Retrieve File Diff Metadata
      description: Fetches a specific file and all of its associated chunks. Used when a user clicks the "wrapper square" in the SVG.
      operationId: getFileDiffMetadata
      tags:
        - Interactive PR
      parameters:
        - $ref: '#/components/parameters/OwnerParam'
        - $ref: '#/components/parameters/RepoParam'
        - $ref: '#/components/parameters/PrNumberParam'
        - $ref: '#/components/parameters/FilenameParam'
      responses:
        '200':
          description: Successfully retrieved the file metadata and its chunks.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileDiffMetadata'
        '404':
          description: File not found in this Pull Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /pr-metadata/chunks:
    get:
      summary: Retrieve Specific Chunk Diff Metadata
      description: Fetches the specific lines of code for a single modified chunk. Used when a user clicks a small green or red square inside the SVG.
      operationId: getChunkDiffMetadata
      tags:
        - Interactive PR
      parameters:
        - $ref: '#/components/parameters/OwnerParam'
        - $ref: '#/components/parameters/RepoParam'
        - $ref: '#/components/parameters/PrNumberParam'
        - $ref: '#/components/parameters/FilenameParam'
        - name: chunkIndex
          in: query
          required: true
          description: The sequential index of the chunk within the file.
          schema:
            type: integer
            example: 0
      responses:
        '200':
          description: Successfully retrieved the chunk metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChunkDiffMetadata'
        '404':
          description: Chunk not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /metrics:
    post:
      summary: Receive metrics using github webhook
      operationId: postMetrics
      security:
        - BearerAuth: [ ]
      tags:
        - Metrics
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/GithubMetrics'
      responses:
        '200':
          description: Metrics processed successfully
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid backendToken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: Get metrics for project
      operationId: getMetrics
      tags:
        - Metrics
      parameters:
        - name: projectUrl
          in: query
          required: true
          description: Project url
          schema:
            type: string
            format: uri
            example: "https://github.com/pjalbrz16/mem.git"
        - name: prId
          in: query
          required: true
          description: Pull request ID
          schema:
            type: string
            example: "123"
      responses:
        '200':
          description: Pull request metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrMetrics'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /listPr:
    get:
      summary: Get list of Pull Requests
      description: Returns a list of PR/MR identification numbers based on the project URL and the detected Git provider.
      tags:
        - RemoteRepository
      parameters:
        - name: projectUrl
          in: query
          description: The full Git clone URL (GitHub, GitLab, Azure, ...)
          required: true
          schema:
            type: string
            format: uri
            example: "https://github.com/pjalbrz16/mem.git"
      responses:
        '200':
          description: A list of PR/MR IDs discovered
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PullRequestInfo'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /addProjectPullRequest:
    post:
      summary: Add a new project for metrics scanning
      tags:
        - RemoteRepository
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProjectMetadata'
      responses:
        '200':
          description: Project added successfully, scanning will start ...
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    githubOAuth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://github.com/login/oauth/authorize
          tokenUrl: https://github.com/login/oauth/access_token
          scopes:
            read:user: Read user profile
            repo: Read repositories
  parameters:
    OwnerParam:
      name: owner
      in: query
      required: true
      description: The GitHub repository owner.
      schema:
        type: string
        example: "octocat"
    RepoParam:
      name: repository
      in: query
      required: true
      description: The name of the GitHub repository.
      schema:
        type: string
        example: "Hello-World"
    PrNumberParam:
      name: pullRequestNumber
      in: query
      required: true
      description: The pull request number/ID.
      schema:
        type: integer
        example: 42
    FilenameParam:
      name: filename
      in: query
      required: true
      description: The full path of the file as recorded in the Git diff.
      schema:
        type: string
        example: "src/main/App.java"
  schemas:
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
    PrMetrics:
      type: object
      properties:
        sonarMetrics:
          $ref: '#/components/schemas/SonarData'
        sarif:
          type: object
          additionalProperties: true
          description: "The full SARIF JSON object (Static Analysis Results Interchange Format)"
        dotFile:
          type: string
          format: graphviz-dot
          description: "The Graphviz .dot string representing data flow"
          example: "digraph G { node1 -> node2 }"
    GithubMetrics:
      type: object
      properties:
        prId:
          type: string
          description: The Pull Request ID
        projectUrl:
          type: string
          description: The URL of the repository scanned
        sarifFile:
          type: string
          format: binary
          description: The CodeQL java.sarif output file
        impactedFiles:
          type: string
          format: binary
          description: The generated impacted_files.csv
        callGraphCsv:
          type: string
          format: binary
          description: The generated call_graph.csv from codeQL query results

    SonarData:
      type: object
      properties:
        qualityGateStatus:
          type: string
          example: "OK"
        analysisDate:
          type: string
          format: date-time
        vulnerabilities:
          type: integer
          example: 10
        securityHotspots:
          type: integer
          example: 5
        codeSmells:
          type: integer
          example: 2
        bugs:
          type: integer
          example: 1

    ProjectMetadata:
      type: object
      properties:
        projectUrl:
          type: string
          format: uri
          example: "https://github.com/org/repo"
        prId:
          type: string
          example: "12345"

    PullRequestInfo:
      type: object
      properties:
        id:
          type: string
          description: "The PR number"
        title:
          type: string
        author:
          type: string
        url:
          type: string
          description: "The link to the PR on GitHub"

    GithubRepo:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        html_url:
          type: string
        description:
          type: string

    ChunkDiffMetadata:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1042
        chunkIndex:
          type: integer
          example: 0
        addedLinesCount:
          type: integer
          example: 5
        deletedLinesCount:
          type: integer
          example: 2
        rawCodeLines:
          type: string
          description: The actual text representation of the diff block.
          example: |
            @@ -1,3 +1,4 @@
            -public class App {
            +public class App implements Runnable {
            +    // Added implementation
             }
      required:
        - id
        - chunkIndex
        - rawCodeLines

    FileDiffMetadata:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 305
        filename:
          type: string
          example: "src/main/App.java"
        status:
          type: string
          enum: [ NEW, DELETED, MODIFIED ]
          example: "MODIFIED"
        chunks:
          type: array
          items:
            $ref: '#/components/schemas/ChunkDiffMetadata'
      required:
        - id
        - filename
        - status
        - chunks